refactor(bin-designer): split paramSlice into composed action sub-factories
