--- src/features/bin-designer/store/slices/paramSlice.ts
@@ -1,789 +1,0 @@
-/**
- * Param slice: bin parameters, scoped updaters, compartments, inserts, wall pattern.
- */
-
-import type { Draft } from 'immer';
-import type {
-  DesignerState,
-  BinParams,
-  BaseConfig,
-  LabelTabConfig,
-  ScoopConfig,
-  WallConfig,
-  OverhangConfig,
-  WallCutout,
-  WallSide,
-  WallPatternConfig,
-  FloorPatternConfig,
-  CutoutConfig,
-  Insert,
-  HandleConfig,
-  HandleSide,
-  HandleWallSide,
-  LidConfig,
-  TextStyleDefaults,
-  TextStyleOverride,
-  WallTextSide,
-  WallTextVerticalAlign,
-} from '../../types';
-import type { ItemEnvelope, ItemStructure } from '@/shared/types/item';
-import type { LabelPlateIconId } from '@/shared/constants/labelPlates';
-import { TEXT_MAX_LENGTH } from '../../types/text';
-import {
-  mergeLipConfig,
-  type LipColorConfig,
-  type TopAccentConfig,
-} from '../../types/featureColors';
-import { DEFAULT_BIN_PARAMS, DEFAULT_FLOOR_PATTERN_CONFIG } from '../../constants';
-import { isErr } from '@/core/result';
-import {
-  carryCompartmentTextsByPosition,
-  isRectangularSelection,
-  mergeCells,
-  splitCompartment,
-} from '../../utils/compartments';
-import { validateCompartmentSizes } from '../../utils/validation';
-import { defaultsForNewDesign, paramsNeedHalfGridMode, pushHistoryEntry } from '../helpers';
-import {
-  MASK_CELLS_PER_UNIT,
-  type CellMask,
-  resizeMask,
-  isAllFilled,
-  validateMask,
-} from '@/shared/utils/cellMask';
-
-type Set = (fn: (state: Draft<DesignerState>) => void) => void;
-type Get = () => DesignerState;
-
-export function createParamSlice(set: Set, get: Get) {
-  return {
-    // Param actions
-    setParam: <K extends keyof BinParams>(key: K, value: BinParams[K]) => {
-      // Guard compartment configuration changes against degenerate cell sizes
-      if (key === 'compartments') {
-        const { params } = get();
-        const newCompartments = value as BinParams['compartments'];
-        const result = validateCompartmentSizes(
-          params.width,
-          params.depth,
-          params.wallThickness,
-          newCompartments.cols,
-          newCompartments.rows,
-          newCompartments.thickness,
-          params.gridUnitMm,
-          params.gridUnitMmY
-        );
-        if (isErr(result)) return;
-      }
-
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params[key] = value;
-        // When the bin footprint grows or shrinks, keep a custom shape mask
-        // aligned to the new dimensions. New cells default to filled so a
-        // resize never silently erases the user's existing shape.
-        if ((key === 'width' || key === 'depth') && state.params.cellMask) {
-          state.params.cellMask = reshapeOrClearMask(
-            state.params.cellMask,
-            state.params.width,
-            state.params.depth
-          );
-        }
-      });
-    },
-
-    setParams: (partial: Partial<BinParams>) => {
-      // Guard compartment configuration changes against degenerate cell sizes
-      if (partial.compartments) {
-        const { params } = get();
-        const width = partial.width ?? params.width;
-        const depth = partial.depth ?? params.depth;
-        const wallThickness = partial.wallThickness ?? params.wallThickness;
-        const gridUnitMm = partial.gridUnitMm ?? params.gridUnitMm;
-        const gridUnitMmY = partial.gridUnitMmY ?? params.gridUnitMmY;
-        const result = validateCompartmentSizes(
-          width,
-          depth,
-          wallThickness,
-          partial.compartments.cols,
-          partial.compartments.rows,
-          partial.compartments.thickness,
-          gridUnitMm,
-          gridUnitMmY
-        );
-        if (isErr(result)) return;
-      }
-
-      set((state) => {
-        pushHistoryEntry(state);
-        Object.assign(state.params, partial);
-        // Keep cellMask aligned with the resulting width/depth. Matters for
-        // the dimension-swap button and share-load, both of which route
-        // through setParams without going via setParam('width'|'depth').
-        if (state.params.cellMask) {
-          state.params.cellMask = reshapeOrClearMask(
-            state.params.cellMask,
-            state.params.width,
-            state.params.depth
-          );
-        }
-      });
-    },
-
-    resetToDefaults: () => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params = { ...defaultsForNewDesign() };
-        // Keep UI toggles in sync with the resolved params: a custom default
-        // may carry fractional dimensions (→ half-grid mode), and defaults
-        // always strip `cellMask` (→ shape editor closed). Without this the
-        // toggles would leak the previous design's state (issue #1752).
-        state.ui.halfGridMode = paramsNeedHalfGridMode(state.params);
-        state.ui.shapeEditorOpen = false;
-      });
-    },
-
-    // Scoped updaters -- merge partial into nested config
-    updateBase: (partial: Partial<BaseConfig>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        Object.assign(state.params.base, partial);
-      });
-    },
-
-    // Non-bin item edits. History/undo is intentionally not wired for non-bin
-    // kinds in this transitional phase — we bump the generation epoch directly
-    // so the preview regenerates via the item path.
-    updateStructure: (partial: Partial<ItemStructure>) => {
-      set((state) => {
-        if (state.itemKind === 'bin' || !state.structure) return;
-        Object.assign(state.structure, partial);
-        state.generation.epoch += 1;
-      });
-    },
-
-    updateEnvelope: (partial: Partial<ItemEnvelope>) => {
-      set((state) => {
-        if (state.itemKind === 'bin' || !state.envelope) return;
-        Object.assign(state.envelope, partial);
-        state.generation.epoch += 1;
-      });
-    },
-
-    updateLabel: (partial: Partial<LabelTabConfig>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        Object.assign(state.params.label, partial);
-      });
-    },
-
-    updateScoop: (partial: Partial<ScoopConfig>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        Object.assign(state.params.scoop, partial);
-      });
-    },
-
-    updateWalls: (partial: Partial<WallConfig>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.walls = { ...state.params.walls, ...partial };
-      });
-    },
-
-    updateOverhang: (partial: Partial<OverhangConfig>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        // Overhang is outward-only; clamp each side to >= 0 so the store never
-        // holds an inverting value (the generator clamps too, defensively).
-        const current = state.params.overhang ?? {
-          left: 0,
-          right: 0,
-          front: 0,
-          back: 0,
-          feet: false,
-        };
-        const next = { ...current, ...partial };
-        // Carry the optional wall taper (#2933) through, clamped outward-only.
-        // Per-side inset is clamped against overhang at generation time.
-        const taper = next.taper
-          ? {
-              ...next.taper,
-              bandHeight: Math.max(0, next.taper.bandHeight),
-              left: Math.max(0, next.taper.left),
-              right: Math.max(0, next.taper.right),
-              front: Math.max(0, next.taper.front),
-              back: Math.max(0, next.taper.back),
-            }
-          : undefined;
-        state.params.overhang = {
-          left: Math.max(0, next.left),
-          right: Math.max(0, next.right),
-          front: Math.max(0, next.front),
-          back: Math.max(0, next.back),
-          feet: next.feet ?? false,
-          ...(next.enabled !== undefined ? { enabled: next.enabled } : {}),
-          ...(taper ? { taper } : {}),
-        };
-      });
-    },
-
-    updateWallSide: (side: WallSide, partial: Partial<WallCutout>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.walls = {
-          ...state.params.walls,
-          [side]: { ...state.params.walls[side], ...partial },
-        };
-      });
-    },
-
-    updateHandles: (partial: Partial<HandleConfig>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.handles = { ...state.params.handles, ...partial };
-      });
-    },
-
-    updateHandleSide: (side: HandleWallSide, partial: Partial<HandleSide>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.handles = {
-          ...state.params.handles,
-          [side]: { ...state.params.handles[side], ...partial },
-        };
-      });
-    },
-
-    // Wall pattern actions
-    updateWallPattern: (partial: Partial<WallPatternConfig>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.wallPattern = { ...state.params.wallPattern, ...partial };
-      });
-    },
-
-    updateFloorPattern: (partial: Partial<FloorPatternConfig>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.floorPattern = {
-          ...DEFAULT_FLOOR_PATTERN_CONFIG,
-          ...state.params.floorPattern,
-          ...partial,
-        };
-      });
-    },
-
-    // Feature color actions
-    updateFeatureColors: (patch: {
-      enabled?: boolean;
-      body?: string;
-      lip?: Partial<LipColorConfig>;
-      labelTab?: string;
-      base?: string;
-      scoop?: string;
-      dividers?: string;
-      text?: string;
-      lid?: string;
-      topAccent?: Partial<TopAccentConfig>;
-    }) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        const current = state.params.featureColors;
-        const { lip: lipPatch, topAccent: topAccentPatch, ...rest } = patch;
-        const nextLip: LipColorConfig = lipPatch
-          ? mergeLipConfig(current.lip, lipPatch)
-          : current.lip;
-        const nextTopAccent: TopAccentConfig = topAccentPatch
-          ? { ...current.topAccent, ...topAccentPatch }
-          : current.topAccent;
-        state.params.featureColors = {
-          ...current,
-          ...rest,
-          lip: nextLip,
-          topAccent: nextTopAccent,
-        };
-        // Multi-color toggle drives the color-tool overlay's visibility; if the
-        // user disables multi-color while a tool is active, the overlay unmounts
-        // but `ui.colorTool` would otherwise stay set, leaving the canvas in
-        // crosshair mode and the eyedropper banner ready to flash back on next
-        // enable. Clear all tool state in lockstep.
-        if (patch.enabled === false) {
-          state.ui.colorTool = null;
-          state.ui.swapFirstZone = null;
-          state.ui.hoveredColorZone = null;
-          state.ui.pickerOverlay = null;
-        }
-      });
-    },
-
-    // Click-lock lid actions
-    updateLid: (partial: Partial<LidConfig>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.lid = { ...state.params.lid, ...partial };
-      });
-    },
-
-    // Cutout configuration actions
-    updateCutoutConfig: (partial: Partial<CutoutConfig>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.cutoutConfig = { ...state.params.cutoutConfig, ...partial };
-      });
-    },
-
-    // Compartment actions
-    setCompartmentGrid: (cols: number, rows: number): number => {
-      const { params } = get();
-      const result = validateCompartmentSizes(
-        params.width,
-        params.depth,
-        params.wallThickness,
-        cols,
-        rows,
-        params.compartments.thickness,
-        params.gridUnitMm,
-        params.gridUnitMmY
-      );
-      if (isErr(result)) return 0;
-
-      // Best-effort carry of labels by position; the rest are counted so the
-      // UI can warn instead of dropping them silently (#2337).
-      const carried = carryCompartmentTextsByPosition(params.compartments, cols, rows);
-      const hasCarried = carried.texts.some((text) => text.length > 0);
-
-      set((state) => {
-        pushHistoryEntry(state);
-        const cells = Array.from({ length: rows * cols }, (_, i) => i);
-        // Divider overrides key on adjacencies the fresh uniform grid no longer
-        // has — drop them. Labels carry by position above where they fit.
-        // Plate-width overrides drop too: the fresh grid's compartment sizes
-        // invalidate the old choices, and auto re-fits per compartment.
-        const {
-          compartmentTexts: _t,
-          labelPlateWidths: _w,
-          labelIcons: _i,
-          dividerOverrides: _o,
-          ...keep
-        } = state.params.compartments;
-        state.params.compartments = {
-          ...keep,
-          cols,
-          rows,
-          cells,
-          ...(hasCarried ? { compartmentTexts: carried.texts } : {}),
-        };
-      });
-      return carried.droppedCount;
-    },
-
-    mergeCells: (cellIndices: readonly number[]) => {
-      if (cellIndices.length < 2) return;
-      const { params } = get();
-      const { cols } = params.compartments;
-
-      // Guard before the history entry so a non-rectangular selection (which the
-      // canonical mergeCells rejects with null) stays a true no-op.
-      if (!isRectangularSelection(cols, cellIndices)) return;
-
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.compartments =
-          mergeCells(state.params.compartments, cellIndices) ?? state.params.compartments;
-      });
-    },
-
-    splitCompartment: (compartmentId: number) => {
-      const { params } = get();
-      // Splitting produces individual cells -- validate full grid is viable
-      const result = validateCompartmentSizes(
-        params.width,
-        params.depth,
-        params.wallThickness,
-        params.compartments.cols,
-        params.compartments.rows,
-        params.compartments.thickness,
-        params.gridUnitMm,
-        params.gridUnitMmY
-      );
-      if (isErr(result)) return;
-
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.compartments = splitCompartment(state.params.compartments, compartmentId);
-      });
-    },
-
-    resetCompartments: () => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.compartments = { ...DEFAULT_BIN_PARAMS.compartments };
-      });
-    },
-
-    setCompartmentText: (compartmentId: number, text: string) => {
-      const { params } = get();
-      const clamped = text.slice(0, TEXT_MAX_LENGTH);
-      const prev = params.compartments.compartmentTexts ?? [];
-      // No-op guard: the input commits on both idle and blur (see
-      // CompartmentTextInput), so an idle-flushed value followed by a blur must
-      // not push a second, identical history entry / regeneration.
-      if ((prev[compartmentId] ?? '') === clamped) return;
-
-      set((state) => {
-        pushHistoryEntry(state);
-        const next = prev.slice();
-        while (next.length <= compartmentId) next.push('');
-        next[compartmentId] = clamped;
-        while (next.length > 0 && next[next.length - 1] === '') next.pop();
-        state.params.compartments = {
-          ...state.params.compartments,
-          ...(next.length > 0 ? { compartmentTexts: next } : { compartmentTexts: undefined }),
-        };
-      });
-    },
-
-    setLabelRowText: (row: number, text: string) => {
-      const { params } = get();
-      const clamped = text.slice(0, TEXT_MAX_LENGTH);
-      const prev = params.label.rowTexts ?? [];
-      // Same idle+blur double-commit guard as setCompartmentText.
-      if ((prev[row] ?? '') === clamped) return;
-
-      set((state) => {
-        pushHistoryEntry(state);
-        const next = prev.slice();
-        while (next.length <= row) next.push('');
-        next[row] = clamped;
-        while (next.length > 0 && next[next.length - 1] === '') next.pop();
-        state.params.label = {
-          ...state.params.label,
-          ...(next.length > 0 ? { rowTexts: next } : { rowTexts: undefined }),
-        };
-      });
-    },
-
-    setCompartmentPlateWidth: (compartmentId: number, widthU: number | null) => {
-      const { params } = get();
-      const prev = params.compartments.labelPlateWidths ?? [];
-      // No-op guard: an unchanged value must not push a history entry or
-      // regeneration. A padded explicit null and an absent slot are the same
-      // auto state, so compare through the ?? null lens.
-      if ((prev[compartmentId] ?? null) === widthU) return;
-
-      set((state) => {
-        pushHistoryEntry(state);
-        const next = prev.slice();
-        while (next.length <= compartmentId) next.push(null);
-        next[compartmentId] = widthU;
-        while (next.length > 0 && next[next.length - 1] === null) next.pop();
-        state.params.compartments = {
-          ...state.params.compartments,
-          // Reset to undefined when every entry is auto — dropped from
-          // persisted JSON on stringify (same convention as compartmentTexts).
-          labelPlateWidths: next.length > 0 ? next : undefined,
-        };
-      });
-    },
-
-    setCompartmentPlateIcon: (compartmentId: number, icon: LabelPlateIconId | null) => {
-      const { params } = get();
-      const prev = params.compartments.labelIcons ?? [];
-      // No-op guard mirrors setCompartmentPlateWidth: unchanged values must
-      // not push history or trigger regeneration.
-      if ((prev[compartmentId] ?? null) === icon) return;
-
-      set((state) => {
-        pushHistoryEntry(state, { affectsGeometry: false });
-        const next = prev.slice();
-        while (next.length <= compartmentId) next.push(null);
-        next[compartmentId] = icon;
-        while (next.length > 0 && next[next.length - 1] === null) next.pop();
-        state.params.compartments = {
-          ...state.params.compartments,
-          // Reset to undefined when no compartment carries an icon — dropped
-          // from persisted JSON on stringify (same convention as
-          // compartmentTexts).
-          labelIcons: next.length > 0 ? next : undefined,
-        };
-      });
-    },
-
-    setCompartmentDividerHeight: (height: number | 'auto') => {
-      const { params } = get();
-      const prev = params.compartments.dividerHeight;
-      // No-op guard: the stepper fires on every tick; an unchanged value must
-      // not push a history entry. Treat undefined and 'auto' as the same state.
-      const prevIsAuto = prev === undefined || prev === 'auto';
-      if (height === 'auto' ? prevIsAuto : prev === height) return;
-
-      set((state) => {
-        pushHistoryEntry(state);
-        if (height === 'auto') {
-          // Omit the field entirely so persisted JSON stays tidy and the bin
-          // shares the full-height cache bucket / cut-path geometry.
-          const { dividerHeight: _drop, ...rest } = state.params.compartments;
-          state.params.compartments = rest;
-        } else {
-          state.params.compartments = { ...state.params.compartments, dividerHeight: height };
-        }
-      });
-    },
-
-    setTextDefaults: (partial: Partial<TextStyleDefaults>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.textDefaults = { ...state.params.textDefaults, ...partial };
-      });
-    },
-
-    setLabelTabTextStyle: (overrides: TextStyleOverride | null) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        if (overrides === null) {
-          const { textStyle: _omit, ...rest } = state.params.label;
-          state.params.label = rest;
-        } else {
-          state.params.label = { ...state.params.label, textStyle: overrides };
-        }
-      });
-    },
-
-    setLidText: (text: string) => {
-      const { params } = get();
-      // Trimmed on store: the worker trims before generating, so persisting
-      // outer whitespace would only create no-op history entries and
-      // regenerations for geometry that can't change.
-      const clamped = text.slice(0, TEXT_MAX_LENGTH).trim();
-      // No-op guard: the input commits on both idle and blur; an unchanged
-      // value must not push a history entry or regeneration.
-      if ((params.surfaceText?.lidText ?? '') === clamped) return;
-
-      set((state) => {
-        pushHistoryEntry(state);
-        const { lidText: _drop, ...rest } = state.params.surfaceText ?? {};
-        const next = {
-          ...rest,
-          ...(clamped !== '' ? { lidText: clamped } : {}),
-        };
-        // Drop the whole key when nothing remains so pre-feature designs
-        // (and cleared text) serialize byte-identically.
-        state.params.surfaceText = Object.keys(next).length > 0 ? next : undefined;
-      });
-    },
-
-    setWallText: (side: WallTextSide, text: string) => {
-      const { params } = get();
-      // Trimmed on store, same rationale as setLidText: the worker trims
-      // before generating, so outer whitespace would only create no-op
-      // history entries and regenerations.
-      const clamped = text.slice(0, TEXT_MAX_LENGTH).trim();
-      // No-op guard: same idle-flush + blur double-commit as setLidText.
-      if ((params.surfaceText?.walls?.[side] ?? '') === clamped) return;
-
-      set((state) => {
-        pushHistoryEntry(state);
-        const { walls: prevWalls, wallAlign, ...rest } = state.params.surfaceText ?? {};
-        const { [side]: _drop, ...otherWalls } = prevWalls ?? {};
-        const walls = {
-          ...otherWalls,
-          ...(clamped !== '' ? { [side]: clamped } : {}),
-        };
-        const hasWalls = Object.keys(walls).length > 0;
-        const next = {
-          ...rest,
-          ...(hasWalls ? { walls } : {}),
-          // Alignment is meaningless without wall text — drop it with the
-          // last wall string so the config can collapse to absent.
-          ...(hasWalls && wallAlign !== undefined ? { wallAlign } : {}),
-        };
-        state.params.surfaceText = Object.keys(next).length > 0 ? next : undefined;
-      });
-    },
-
-    clearWallText: () => {
-      const { params } = get();
-      // No-op guard: nothing to clear keeps undo/regeneration quiet. Checks
-      // `wallAlign` too — a programmatic/legacy state can hold an alignment
-      // without any wall strings, and clearing must still drop it.
-      if (params.surfaceText?.walls === undefined && params.surfaceText?.wallAlign === undefined) {
-        return;
-      }
-
-      set((state) => {
-        pushHistoryEntry(state);
-        // Drop `walls` AND `wallAlign` (meaningless without wall text); keep any
-        // other surface-text keys (lid text, shared style) intact.
-        const { walls: _walls, wallAlign: _align, ...rest } = state.params.surfaceText ?? {};
-        state.params.surfaceText = Object.keys(rest).length > 0 ? rest : undefined;
-      });
-    },
-
-    setWallTextAlign: (align: WallTextVerticalAlign) => {
-      const { params } = get();
-      // 'center' is the implicit default — stored as an absent field.
-      const normalized = align === 'center' ? undefined : align;
-      if ((params.surfaceText?.wallAlign ?? undefined) === normalized) return;
-
-      set((state) => {
-        pushHistoryEntry(state);
-        const { wallAlign: _drop, ...rest } = state.params.surfaceText ?? {};
-        const next = {
-          ...rest,
-          ...(normalized !== undefined ? { wallAlign: normalized } : {}),
-        };
-        state.params.surfaceText = Object.keys(next).length > 0 ? next : undefined;
-      });
-    },
-
-    setSurfaceTextStyle: (overrides: TextStyleOverride | null) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        const { style: _drop, ...rest } = state.params.surfaceText ?? {};
-        const next = {
-          ...rest,
-          ...(overrides !== null && Object.keys(overrides).length > 0 ? { style: overrides } : {}),
-        };
-        state.params.surfaceText = Object.keys(next).length > 0 ? next : undefined;
-      });
-    },
-
-    setDividerOverride: (
-      compartmentA: number,
-      compartmentB: number,
-      offsetStart: number,
-      offsetEnd: number
-    ) => {
-      // Enforce canonical pair ordering: the validator + worker lookup all
-      // assume compartmentA < compartmentB, and silently allowing unordered
-      // pairs at the store would let two storage representations of the
-      // same divider exist as separate entries.
-      const [a, b] =
-        compartmentA < compartmentB ? [compartmentA, compartmentB] : [compartmentB, compartmentA];
-      const { params } = get();
-      const prev = params.compartments.dividerOverrides ?? [];
-      const existing = prev.find((o) => o.compartmentA === a && o.compartmentB === b);
-      // No-op guard: dragging an endpoint to its current position fires this
-      // action; an unchanged value would otherwise push a history entry per
-      // pointer move and bloat the undo stack.
-      if (existing && existing.offsetStart === offsetStart && existing.offsetEnd === offsetEnd) {
-        return;
-      }
-      set((state) => {
-        pushHistoryEntry(state);
-        const next = prev.filter((o) => !(o.compartmentA === a && o.compartmentB === b));
-        // Treat zero offsets as "remove" so the storage stays tidy and the
-        // empty array can be omitted from persisted JSON.
-        if (offsetStart !== 0 || offsetEnd !== 0) {
-          next.push({ compartmentA: a, compartmentB: b, offsetStart, offsetEnd });
-        }
-        state.params.compartments = {
-          ...state.params.compartments,
-          ...(next.length > 0 ? { dividerOverrides: next } : { dividerOverrides: undefined }),
-        };
-      });
-    },
-
-    removeDividerOverride: (compartmentA: number, compartmentB: number) => {
-      const [a, b] =
-        compartmentA < compartmentB ? [compartmentA, compartmentB] : [compartmentB, compartmentA];
-      const { params } = get();
-      const prev = params.compartments.dividerOverrides ?? [];
-      const next = prev.filter((o) => !(o.compartmentA === a && o.compartmentB === b));
-      if (next.length === prev.length) return; // nothing changed
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.compartments = {
-          ...state.params.compartments,
-          ...(next.length > 0 ? { dividerOverrides: next } : { dividerOverrides: undefined }),
-        };
-      });
-    },
-
-    clearDividerOverrides: () => {
-      const { params } = get();
-      if (!params.compartments.dividerOverrides?.length) return;
-      set((state) => {
-        pushHistoryEntry(state);
-        const { dividerOverrides: _drop, ...rest } = state.params.compartments;
-        state.params.compartments = rest;
-      });
-    },
-
-    // Insert actions
-    addInsert: (insert: Insert) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.inserts = [...state.params.inserts, insert];
-      });
-    },
-
-    removeInsert: (id: string) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.inserts = state.params.inserts.filter((i) => i.id !== id);
-      });
-    },
-
-    updateInsert: (id: string, updates: Partial<Insert>) => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.inserts = state.params.inserts.map((i) =>
-          i.id === id ? { ...i, ...updates } : i
-        );
-      });
-    },
-
-    clearInserts: () => {
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.inserts = [];
-      });
-    },
-
-    // Custom bin shape (cellMask). Setting undefined or a fully-filled mask
-    // routes the generator through the rectangle fast-path. Partial masks
-    // produce a polygon footprint. Rejects masks that fail structural
-    // validation (empty / disconnected / holes) or whose dimensions don't
-    // match the current width/depth at half-bin resolution — a mismatched
-    // mask would otherwise trip assertValidMask in the generator.
-    setCellMask: (mask: CellMask | undefined) => {
-      let next: CellMask | undefined;
-      if (mask === undefined || isAllFilled(mask)) {
-        next = undefined;
-      } else {
-        const { width, depth } = get().params;
-        if (mask.cols !== Math.round(width * MASK_CELLS_PER_UNIT)) return;
-        if (mask.rows !== Math.round(depth * MASK_CELLS_PER_UNIT)) return;
-        if (validateMask(mask) !== null) return;
-        next = mask;
-      }
-      set((state) => {
-        pushHistoryEntry(state);
-        state.params.cellMask = next;
-      });
-    },
-  };
-}
-
-/**
- * Resize a cellMask to match new `width × depth` (in grid units). If the
- * resized mask turns out to be structurally invalid (very rare — the caller
- * changed dimensions in a way that disconnects the shape) or if it now
- * covers the full footprint, return `undefined` so the generator drops back
- * to the rectangle fast-path.
- */
-function reshapeOrClearMask(
-  mask: CellMask,
-  widthUnits: number,
-  depthUnits: number
-): CellMask | undefined {
-  const cols = Math.round(widthUnits * MASK_CELLS_PER_UNIT);
-  const rows = Math.round(depthUnits * MASK_CELLS_PER_UNIT);
-  if (mask.cols === cols && mask.rows === rows) return mask;
-  const resized = resizeMask(mask, cols, rows);
-  if (isAllFilled(resized)) return undefined;
-  if (validateMask(resized) !== null) return undefined;
-  return resized;
-}
